% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBayes_function.R
\name{plotBayes}
\alias{plotBayes}
\title{Function for exploring Bayesian inference for the sample mean for normal data.}
\usage{
plotBayes(data, prior.mean, prior.sd, min, max, points = 1000)
}
\arguments{
\item{data}{A vector of data.}

\item{prior.mean}{The mean of the prior distribution for the mean parameter mu.}

\item{prior.sd}{The SD of the prior distribution for the mean. Smaller means more certainty.}

\item{min}{The minimum possible value of mu to consider.}

\item{max}{The maximum possible value of mu to consider.}

\item{points}{The number of values of mu to be calculated, higher means more precision. Defaults to 1000.}
}
\description{
This function plots the prior, likelihood, and posterior distribution of the sample mean. It reports the MAP (maximum a posteriori) estimate of the mean as well as the 95\% credible interval.
}
\examples{
set.seed(1)
data <- rnorm(n=10, mean=1, sd=1)
plotBayes(data, prior.mean=0, prior.sd=.2, min=-2, max=2)
}
