% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBayes_function.R
\name{plotBayes}
\alias{plotBayes}
\title{Function for exploring Bayesian inference for the sample mean for continuous data.}
\usage{
plotBayes(data, prior.type = "normal", prior.parameters, min, max,
  points = 1000)
}
\arguments{
\item{data}{A vector of data}

\item{prior.type}{The type of distribution for the prior. Can be "uniform" or "normal"}

\item{prior.parameters}{A vector of length two providing the parameter values for the prior. When prior.type="normal", the values are the mean and sd. When prior.type="uniform", the values are the min and the max.}

\item{min}{The minimum possible value of mu to consider.}

\item{max}{The maximum possible value of mu to consider.}

\item{points}{The number of values of mu to be calculated, higher means more precision. Defaults to 1000.}
}
\description{
This function plots the prior, likelihood, and posterior distribution of the sample mean. It reports the MAP (maximum a posteriori) estimate of the mean as well as the 95\% credible interval.
}
\examples{
set.seed(1)
data <- rnorm(n=10, mean=1, sd=1)
plotBayes(data, prior.type="normal", prior.parameters=c(0, .5), min=-2, max=2) 
plotBayes(data, prior.type="uniform", prior.parameters=c(.7, 1.5), min=-2, max=2)
}
